Share await_consensus_sync function
